1.2.0

+++ lib/get_observer_add.dart
@@ -40,8 +40,8 @@ class GlobalRouteObserver extends GetObserver {
 
     var loginApi = await ApiClass().accessPay();
     if (loginApi != null && loginApi['success'] == true) {
-      if (loginApi['pay_by_account'] != null) {
-        if (loginApi['pay_by_account'] == 'no') {
+      if (loginApi['has_app_access'] != null) {
+        if (loginApi['has_app_access'] == 'no') {
           final prefs = await SharedPreferences.getInstance();
           await prefs.clear();
 
@@ -55,7 +55,7 @@ class GlobalRouteObserver extends GetObserver {
               builder: (BuildContext context) {
                 return AlertDialog(
                   title: const Text(
-                    "Sorry, you can not complete the login",
+                    "Sorry, you can not continue.",
                     style: TextStyle(
                       color: Colors.black87,
                       fontSize: 18,
@@ -63,7 +63,7 @@ class GlobalRouteObserver extends GetObserver {
                     ),
                   ),
                   content: const Text(
-                    "Your pay by account is disable.",
+                    "Your account is deactivate.",
                     style: TextStyle(
                       color: Colors.black87,
                       fontSize: 16,
@@ -89,12 +89,69 @@ class GlobalRouteObserver extends GetObserver {
               },
             );
           });
+        } else {
+          if (loginApi['pay_by_account'] != null) {
+            if (loginApi['pay_by_account'] == 'no') {
+              final prefs = await SharedPreferences.getInstance();
+              await prefs.clear();
+
+              Get.offAll(() => LoginPage(), routeName: '/LoginPage');
+
+              // Show dialog AFTER pushing login page
+              Future.delayed(Duration(milliseconds: 100), () {
+                showDialog<void>(
+                  context: Get.context!,
+                  barrierDismissible: false,
+                  builder: (BuildContext context) {
+                    return AlertDialog(
+                      title: const Text(
+                        "Sorry, you can not continue.",
+                        style: TextStyle(
+                          color: Colors.black87,
+                          fontSize: 18,
+                          fontWeight: FontWeight.w800,
+                        ),
+                      ),
+                      content: const Text(
+                        "Your pay by account is disable.",
+                        style: TextStyle(
+                          color: Colors.black87,
+                          fontSize: 16,
+                          fontWeight: FontWeight.w600,
+                        ),
+                      ),
+                      actions: [
+                        ElevatedButton(
+                          style: ElevatedButton.styleFrom(
+                            backgroundColor: AppColors.nakedSyrup,
+                            padding: const EdgeInsets.all(8),
+                          ),
+                          child: const Text(
+                            "Close",
+                            style: TextStyle(color: Colors.white, fontSize: 14),
+                          ),
+                          onPressed: () {
+                            Get.back();
+                          },
+                        ),
+                      ],
+                    );
+                  },
+                );
+              });
+            }
+          } else {
+            final prefs = await SharedPreferences.getInstance();
+            await prefs.clear();
+            Get.offAll(() => LoginPage(), routeName: '/LoginPage');
+          }
         }
-      } else {
-        final prefs = await SharedPreferences.getInstance();
-        await prefs.clear();
-        Get.offAll(() => LoginPage(), routeName: '/LoginPage');
       }
+      // else {
+      //   final prefs = await SharedPreferences.getInstance();
+      //   await prefs.clear();
+      //   Get.offAll(() => LoginPage(), routeName: '/LoginPage');
+      // }
     }
   }
 }
